backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
tensorflow>=2.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
geopy>=2.3.0
google-generativeai>=0.1.0
python-multipart>=0.0.6
torch>=2.0.0
torchvision>=0.15.0
firebase-admin>=6.5.0
